sql: fix integer overflow in built-in functions
This patch replaces the type for some int and uint32_t values with
size_t to avoid problems with integer overflow.

Closes tarantool/security#119

NO_DOC=bugfix

(cherry picked from commit 60a187d7e113fb1c0768774d46215792cc0cef7e)

+++ src/box/sql/sqlInt.h
@@ -2429,7 +2429,13 @@ void *sqlDbReallocOrFree(sql *, void *, u64);
 void *sqlDbRealloc(sql *, void *, u64);
 void sqlDbFree(sql *, void *);
 int sqlMallocSize(void *);
-int sqlDbMallocSize(sql *, void *);
+
+/**
+ * Return the size of a memory allocation previously obtained from
+ * sqlDbMallocRaw() or sqlDbMallocRawNN().
+ */
+size_t
+sqlDbMallocSize(sql *, void *);
 
 /*
  * On systems with ample stack space and that support alloca(), make
